Updated Motion Detector Texture and Animation
I retextured the motion detector and laser array to appear more reflective, and less harsh when reflecting a lot of light using shading nodes. I also added an animation zooming into the laser array as practice for an animation of the laser array zooming across the screen when triggered. In addition I worked a small bit on designing the actual lasers, but I could not yet make them emit light.