Add files via upload
This is a prototype commit of a simple rusting texture. It is great for flat surfaces, but the texture changes when a smooth surface is moved. It was inspired by a YouTube tutorial that shows how to make striped materials, but the rust design, color palette, and texture were all original. The grainy style of the material fits perfectly inside Lethal Company's low-poly environment, and it is encouraged to be used for any other design that might need a rusty aesthetic. The YouTube tutorial used is linked below.


https://www.youtube.com/watch?v=jrLwxf4CWfo