"Double Stripe" Texture
This is a texture that will create stripes on a surface. Works best when viewed from a head-on angle (see render). Works great for flat and round objects. Did not encounter the same error as in the previous texture.

The texture's main conception is a wave node connected to a color ramp. The wave node can be used to modify stripe dimensions, phase offset, direction, and uniformity. The color node is used to add colors to the stripes. You need at least two colors present on the ramp for stripes to appear. Add more colors by adding more sliders. To add colored stripes on a single-colored background separate each color chevron with one corresponding to the background color. See the .blend file's original shader setup for further description. It utilizes a double stripe pattern of pink and green on a black surface, but in reality it is a black stripe + pink stripe + black stripe + green stripe. Happy editing!