Flaming Candle
This is a candle commit that will be used in a birthday cake. I played around with using the light to give the illusion of a flame's brightness. It appears quite powerful in the render, so I plan to tone it down. It also uses a custom Blue/white texture based off of the commit I made to my textures branch. The Candle was optimized to be low-poly, so I hope it can be used in other aspects of the game.