Cake commit
This commit is a birthday cake that features the candles from the previous commit in this branch. The blend file implements the scheme in cakeWithFlame, which has fine-tuned the light idea explored in the candle commit. Additionally, this design features many concepts learned in the donut video, such as snap to face and proportional editing.